requirements.txt
apache-airflow==2.5.3
pandas>=2.0.1
requests==2.28.2
python-dotenv
boto3
google-cloud-bigquery
jsonformatter
pyarrow
apache-airflow[cncf.kubernetes]